chore(editor): move memory address resolving into editor-state
