+++ packages/editor/packages/editor-state/src/features/binary-assets/effect.ts
@@ -1,6 +1,7 @@
 import type { State } from '~/types';
 
 import { EventDispatcher } from '~/types';
+import resolveBinaryAssetTarget from '~/pureHelpers/resolveBinaryAssetTarget';
 
 export default function binaryAssets(state: State, events: EventDispatcher): () => void {
 	async function onLoadBinaryFilesIntoMemory() {
@@ -11,7 +12,19 @@ export default function binaryAssets(state: State, events: EventDispatcher): () 
 
 		const assets = state.compiledConfig.binaryAssets || [];
 		for (const asset of assets) {
-			await state.callbacks.loadBinaryFileIntoMemory(asset);
+			const resolved = resolveBinaryAssetTarget(state, asset.memoryId);
+			if (!resolved) {
+				console.warn('Unable to resolve memory target:', asset.memoryId);
+				continue;
+			}
+
+			await state.callbacks.loadBinaryFileIntoMemory({
+				url: asset.url,
+				moduleId: resolved.moduleId,
+				memoryName: resolved.memoryName,
+				byteAddress: resolved.byteAddress,
+				byteLength: resolved.byteLength,
+			});
 		}
 	}
