+++ packages/editor/packages/editor-state/src/index.ts
@@ -107,6 +107,7 @@ export type {
 	GraphicHelper,
 	BinaryAsset,
 	ConfigBinaryAsset,
+	ResolvedBinaryAsset,
 	Callbacks,
 	ExampleModule,
 	ModuleMetadata,
