+++ packages/editor/src/binaryAssets/clearBinaryAssetCache.ts
@@ -1,0 +1,9 @@
+import { BINARY_ASSET_CACHE_NAME } from './constants';
+
+export default async function clearBinaryAssetCache(): Promise<void> {
+	if (typeof caches === 'undefined') {
+		return;
+	}
+
+	await caches.delete(BINARY_ASSET_CACHE_NAME);
+}
