+++ packages/editor/src/binaryAssets/constants.ts
@@ -1,0 +1,1 @@
+export const BINARY_ASSET_CACHE_NAME = '8f4e-binary-assets';
